requirements.txt
azure-identity
azure-mgmt-costmanagement
azure-mgmt-resource
